Matching computation for ADD-POINT, flatness detection in COMPUTE-INITIAL-MESH
Before this change, ADD-POINT used

  dist := (plane-sigdist vertices vertex plane)
  (> dist (* eps |plane|)) phrased as (> dist^2 (* eps^2 |plane|^2))

while COMPUTE-INITIAL-MESH USED

  dist := (plane-sigdist vertices vertex plane)
  (> dist eps)

After this change,

  (> dist^2 (* eps^2 |plane|^2))

is used in both places.

The problem caused by the previous version is as follows: consider a
set of points that almost lie in a plane such as

  #(0.38144001364707947d0 0.2354149967432022d0 -0.23686400055885315d0
    0.7900480031967163d0 0.23541399836540222d0 1.4066979885101318d0
    2.5108859539031982d0 0.23541000485420227d0 0.9544129967689514d0
    2.26465106010437d0 0.2354149967432022d0 -0.33437100052833557d0
    2.26465106010437d0 0.2354149967432022d0 -0.9228739738464355d0)

The previous version of the flatness detection in COMPUTE-INITIAL-MESH
would classify this set of points as not flat. However, ADD-POINT
would not consider one of the to lie "inside" any of the generated
initial planes due the different distance comparison.